feat: create caddy module for lura
At this point the module instantiates a static lura gateway. That should
be fixed in following commits.

The lura router layer is expected to run a server. Once the server
itself is run by caddy, we provided a custom run function that only
captures the generated lura http handler.

It is expected that a custom router layer should be specifically
designed for this integration with caddy.