feat: allow endpoint configuration in Caddyfile
The lura api gateway configuration was statically defined in source
code, which is not ideal for any environments other than testing.

Implement Caddyfile unmarshalling so that basic configuration of lura
might be customizable.

At this point not all lura configurations was made available, but only
the bare minimum for simple endpoint definitions.