feat: write a custom router layer
The default router layer is based on the gin framework, which is
overkill for the use case of an embedded lura instance.

Also, the gin adapter provided by lura does not the same level of
customization that the mux router provides.

Write a new router layer based on the standard mux one with httprouter.

The proxy stack was also customized, so that caddy placeholders might be
used when defining the backend url pattern.