refactor: rewrite the router layer
The router layer abstraction provided by lura is expected to be executed
as a standalone application, which is not the case in this integration,
what forced us to implement workarounds in order to access the inner
http handler.

Also, as the errors generated by the proxy layer are not propagated
outside the router, because the router implements the default `http.Handler`,
it was not possible to propagate these erros through caddy global error
handling mechanism.

So it seems to make sense to write from scratch a router that better
integrates with caddy.

The http router was copied from the `julienschmidt/httprouter` and
refactored to work with `caddyhttp.Handler` interface.

Part of the mux router layer implementation provided by the lura team
was also copied and modified accordingly.

Regarding the usage of placeholders in backend request, the processing
was done partially by caddy and partially by lura itself, now every url
path parameter is exposed as caddy placeholder and replaced by using a
`caddy.Replacer` instance.